pyproject.toml
[project]
name = "kagyucalendarapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Sergei Bykov",email = "ss.bykoff@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "fastapi (>=0.115.7,<0.116.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "fastapi-users[sqlalchemy] (>=14.0.1,<15.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "fake-headers (>=1.0.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"
black = "^25.1.0"

